port extend sectors test to run on the dyn VM
